session created

+++ app/src/main/java/com/example/swift/businessLayer/dataClasses/Driver.kt
@@ -7,7 +7,7 @@ class Driver (
     override var age:String,
     override var gender:String,
     override var phoneNumber:String,
-    override var rating:Float,
+    override var rating: Double,
     var cnic:String,
     var licenseNumber:String,
     var vehicleType:String,
